backend_service/pyproject.toml
[project]
name = "modal-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "nheingit", email = "60185486+nheingit@users.noreply.github.com" }
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.6",
    "google-api-python-client>=2.160.0",
    "google-auth>=2.38.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "modal>=0.72.7",
    "openai>=1.61.0",
    "python-dotenv>=1.0.1",
]

[project.scripts]
modal-app = "modal_app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
